+++ crates/scheduler/src/applet/store.rs
@@ -61,14 +61,16 @@ pub trait MemoryApi {
         Ok(bytemuck::from_bytes_mut(self.get_mut(ptr, core::mem::size_of::<T>() as u32)?))
     }
 
-    fn alloc_copy(&mut self, ptr_ptr: u32, len_ptr: Option<u32>, data: &[u8]) -> Result<(), Trap> {
+    fn alloc_copy(&mut self, ptr_ptr: u32, len_ptr: Option<u32>, data: &[u8]) -> Result<u32, Trap> {
         let len = data.len() as u32;
-        let ptr = self.alloc(len, 1)?;
-        self.get_mut(ptr, len)?.copy_from_slice(data);
-        self.get_mut(ptr_ptr, 4)?.copy_from_slice(&ptr.to_le_bytes());
+        if 0 < len {
+            let ptr = self.alloc(len, 1)?;
+            self.get_mut(ptr, len)?.copy_from_slice(data);
+            self.get_mut(ptr_ptr, 4)?.copy_from_slice(&ptr.to_le_bytes());
+        }
         if let Some(len_ptr) = len_ptr {
             self.get_mut(len_ptr, 4)?.copy_from_slice(&len.to_le_bytes());
         }
-        Ok(())
+        Ok(len)
     }
 }