+++ crates/api-desc/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Minor
 
+- Improve safety documentation
+- Add `Api::wasm_markdown()` for top-level documentation
 - Use Rust edition 2024
 - Use C-string literals to implement dispatch for native applets
 - Implement `bytemuck::Pod` for generated `Params`
@@ -137,4 +139,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 1 -->
+<!-- Increment to skip CHANGELOG.md test: 2 -->